snd_mp3tag: avoid signed-shift UB in the little-endian byte assembly
`(data[23] << 24)` promotes the unsigned char to int, so a top-bit-set
byte overflows a signed int -- undefined behavior. gcc and clang happen to
produce the wraparound the code wants, which is why this never showed up
on Linux or macOS, but MSVC is free to assume the shift does not overflow
and fold `(x << 24) & 0x80000000` to zero.

That is what it does: with the Windows differential suite finally linking,
`mp3_skiptags_differential` failed on the `ape2-end` case with the C
oracle skipping 132 bytes where the Rust port skipped 164 -- exactly the
32-byte APE header, i.e. the `flags & (1U << 31)` header-present test in
get_ape_len silently evaluating false. Left alone, MSVC builds hand 32
bytes of tag to the mp3 decoder on every APEv2 file that carries a header.

Cast each byte to unsigned before shifting, at all seven sites in the
file. The values are bit-identical to what gcc/clang produce today: the
two `(int)` sites and the `(long)(int)` in get_ape_len keep the same
truncation, and the `unsigned` sites were already being cast to unsigned.
This also makes the C match the Rust port's `i32/u32::from_le_bytes`
exactly, which is what the differential asserts.

+++ Quake/snd_mp3tag.c
@@ -84,7 +84,7 @@ static qboolean is_apetag (const unsigned char *data, size_t length)
 	{
 		return false;
 	}
-	v = (unsigned)((data[11] << 24) | (data[10] << 16) | (data[9] << 8) | data[8]); /* version */
+	v = ((unsigned)data[11] << 24) | ((unsigned)data[10] << 16) | ((unsigned)data[9] << 8) | (unsigned)data[8]; /* version */
 	if (v != 2000U && v != 1000U)
 	{
 		return false;
@@ -99,9 +99,9 @@ static qboolean is_apetag (const unsigned char *data, size_t length)
 static long get_ape_len (const unsigned char *data)
 {
 	unsigned int flags, version;
-	long		 size = (long)((data[15] << 24) | (data[14] << 16) | (data[13] << 8) | data[12]);
-	version = (unsigned)((data[11] << 24) | (data[10] << 16) | (data[9] << 8) | data[8]);
-	flags = (unsigned)((data[23] << 24) | (data[22] << 16) | (data[21] << 8) | data[20]);
+	long		 size = (long)(int)(((unsigned)data[15] << 24) | ((unsigned)data[14] << 16) | ((unsigned)data[13] << 8) | (unsigned)data[12]);
+	version = ((unsigned)data[11] << 24) | ((unsigned)data[10] << 16) | ((unsigned)data[9] << 8) | (unsigned)data[8];
+	flags = ((unsigned)data[23] << 24) | ((unsigned)data[22] << 16) | ((unsigned)data[21] << 8) | (unsigned)data[20];
 	if (version == 2000U && (flags & (1U << 31)))
 		size += 32; /* header present. */
 	return size;
@@ -214,8 +214,8 @@ static long get_musicmatch_len (snd_stream_t *stream)
 
 	FS_fseek (&stream->fh, -68, SEEK_END);
 	FS_fread (buf, 1, 20, &stream->fh);
-	imgext_ofs = (int)((buf[3] << 24) | (buf[2] << 16) | (buf[1] << 8) | buf[0]);
-	version_ofs = (int)((buf[15] << 24) | (buf[14] << 16) | (buf[13] << 8) | buf[12]);
+	imgext_ofs = (int)(((unsigned)buf[3] << 24) | ((unsigned)buf[2] << 16) | ((unsigned)buf[1] << 8) | (unsigned)buf[0]);
+	version_ofs = (int)(((unsigned)buf[15] << 24) | ((unsigned)buf[14] << 16) | ((unsigned)buf[13] << 8) | (unsigned)buf[12]);
 	if (version_ofs <= imgext_ofs)
 		return -1;
 	if (version_ofs <= 0 || imgext_ofs <= 0)
@@ -263,7 +263,7 @@ static long get_musicmatch_len (snd_stream_t *stream)
 		return -1;
 	FS_fseek (&stream->fh, -len, SEEK_END);
 	FS_fread (buf, 1, 8, &stream->fh);
-	j = (int)((buf[7] << 24) | (buf[6] << 16) | (buf[5] << 8) | buf[4]);
+	j = (int)(((unsigned)buf[7] << 24) | ((unsigned)buf[6] << 16) | ((unsigned)buf[5] << 8) | (unsigned)buf[4]);
 	if (j < 0)
 		return -1;
 	/* verify image size: */
